Replace ol/Sphere with ol/sphere
