+++ examples/tissot.js
@@ -1,6 +1,5 @@
 import _ol_Feature_ from '../src/ol/Feature.js';
 import _ol_Map_ from '../src/ol/Map.js';
-import _ol_Sphere_ from '../src/ol/Sphere.js';
 import _ol_View_ from '../src/ol/View.js';
 import Polygon from '../src/ol/geom/Polygon.js';
 import TileLayer from '../src/ol/layer/Tile.js';
@@ -57,13 +56,11 @@ var map3857 = new _ol_Map_({
   })
 });
 
-var wgs84Sphere = new _ol_Sphere_(6378137);
-
 var radius = 800000;
 var x, y;
 for (x = -180; x < 180; x += 30) {
   for (y = -90; y < 90; y += 30) {
-    var circle4326 = Polygon.circular(wgs84Sphere, [x, y], radius, 64);
+    var circle4326 = Polygon.circular([x, y], radius, 64);
     var circle3857 = circle4326.clone().transform('EPSG:4326', 'EPSG:3857');
     vectorLayer4326.getSource().addFeature(new _ol_Feature_(circle4326));
     vectorLayer3857.getSource().addFeature(new _ol_Feature_(circle3857));
